+++ packages/authhero/test/routes/auth-api/token.spec.ts
@@ -286,6 +286,7 @@ describe("token", () => {
           code_id: "123456",
           login_id: loginSesssion.id,
           expires_at: new Date(Date.now() + 1000 * 60 * 5).toISOString(),
+          redirect_uri: "http://example/callback",
         });
 
         const response = await client.oauth.token.$post(